add stab-clause rule
This should prepare the grammar to take on functions, which are
currently commented out in this commit. The documentation on
function definitions:

    Name(Pattern11,...,Pattern1N) [when GuardSeq1] ->
        Body1;
    ...;
    Name(PatternK1,...,PatternKN) [when GuardSeqK] ->
        BodyK.

Looks quite similar to anonymous functions:

    fun
        [Name](Pattern11,...,Pattern1N) [when GuardSeq1] ->
                  Body1;
        ...;
        [Name](PatternK1,...,PatternKN) [when GuardSeqK] ->
                  BodyK
    end

Right? I think we should be able to reuse the _named_stab_clause
rule between the two, and will probably be able to re-use it again
for specs.