allow newline to be a terminator of statements
This is very much against the grain of the original parser as
far as I can tell, but it allows the grammar to be much more permissive
about how expressions can be written.

See the test case: the second (call) would be mangled without this
choice of newline because of the precedences of (binary_operator) and
(call).