parse function terminators more strictly on '.'
Allowing '\n' to be a terminator has problems when a binary expression
is spread out over a line, as in the test case. The '\n' is consumed as
the terminator despite the binary operator being a longer pattern. In
this case, the '++' is parsed as two unary operators. There might be a
way to fix this without disallowing '\n' as a terminator for functions
(maybe by preventing double unary operators, although that has problems
of its own). However the test cases don't show any interest in using
that behavior: it's fine to be strict about it since this change is only
for function definitions (and it's more correct syntactically).