allow any printable character to be a '$c' character
Unicode and such are allowed in characters, so it's easier to specify
what isn't allowed: escape sequences (which are then matched in the
choice).

This change is a little cause for celebration: the grammar now
parses 100% of the OTP compiler .erl files. More testing is needed
and more bugs to be found in the other .erl/.hrl files of the OTP
code-base, but this is not small thing! 🎊