allow writing function clauses via macros
This is done in OTP in ssh.hrl and some consumers of the header.
It's somewhat making me regret the decision to have function clauses
not be '_expression's, but not enough to try to poke that hornet's
nest yet. I'll need to see some more examples of funky macros to be
satisfied.