allow bodies to macro definitions
The 'define' attribute can have a number of "arguments" as they
were previously parsed. In fact it can only have two arguments:
the macro name followed by the body that will be spliced into
wherever the macro is used. Now the rules reflect that: you can
have the definition _expression and then a _body.