coalesce 'quoted_atom' rule into 'atom'
The 'atom' node is sufficient, we don't need to differentiate the
nodes based on whether or not they're quoted. It just complicates
the highlights so that anywhere where we would want to match an
'atom', we instead need to use the more complicated pattern

    [(atom) (quoted_atom)]

which is just a hassle. I would imagine it's theoretically faster
to just use an 'atom' rule at query time as well, and queries are
usually the slow performance part of tree-sitter in my experience.