fix remote call confusion with binary operators
Some cases exhibited strange behavior for remote calls without terminators:

    erlang:nif_error(undefined)

Without a terminator (or a newline acting as a sort of terminator), this
would be recognized as a binary operator rather than a call, confusingly.

This fix is not really ideal, it dumbly removes `$.call` from the set of
valid right-hand-sides of the `:` binary operator. This looks to be valid
looking at the bitstring syntax[^1] but is nonetheless hacky.

[^1]: https://www.erlang.org/doc/reference_manual/expressions.html#bit-syntax-expressions