Fix parsing of multiline funs in larger expressions
This is a fix for some complicated interactions between anonymous
functions, macros, and the `->` operator.

Multi-line `fun`s in larger expressions would routinely allow the
`->` operator to consume the `fun(Args)` and the first expression
in the body of the `fun`. This was almost unnoticeable because the
remainder of the function body would parse as valid binary operators
within a larger `body` node. The `end` token - a keyword - would
be highlighted as an atom which was the only visual indication.

The fix was to remove `->` as a valid binary operator. It should
only be allowed in function declarations. To fix this though, the
`_named_stab_clause` rule needed to eliminate the possibility of
containing a macro with arguments. A snippet like

    ?MODULE() -> bar.

is a valid function declaration. It is tough to parse because the
arguments - `()` - belong to the function declaration instead of
the macro. This is hard to express precedence-wise. Through some
odd tree-sitter precedence rules, it works to split the possible
macro constant values and list them explicitly.

Note that **it is not possible** to replace the split macro rules
with the combined `macro` node.

Included is a refactor which splits the `macro` rule into two
unnamed rules: `_macro_constant` and `_macro_call`. Usage of
one but not both of these is a potential indicator that valid
syntax is being disallowed. For trade-offs like this case though,
the parser is better off overall.