Support currently implemented sigils
At time of writing we have the default sigil (no prefix), 's', 'b', 'S'
and 'B'. The way that the sigil is parsed depends on the sigil's prefix,
so we can't generically parse all future sigils.

For example the 'r' sigil for regular expressions is proposed to read
the contents verbatim even though it's lowercase, breaking with the
pattern defined in Elixir and carried forward with 's' and 'b'.

So for now we follow upstream and only parse what they have implemented.